web_gateway: carve settings.rs (pure move)
The settings surface moves to web_gateway/settings.rs: SettingsPayload
(de)serialization + normalization, runtime overrides, POST application
+ control-msg dispatch, API-key set/status, and the thin per-route
stream handlers, with their 6 tests.
